Added Augment & Imbuement Smart Protections
Blocking augment of stackables, charged items, non moveables ect.
Checking that item is in the slot its assigned to for the augments to count if the config is set to true (default), only a problem for classic slots.
Preventing items from decaying and/or transforming if they are imbued or augmented.
Imbuement stackables still lack necessary support and protection for unwanted behavior.